Added a Run subclass for Ind race types.
