+++ src/core/IndRun.java
@@ -1,0 +1,57 @@
+package core;
+
+public class IndRun extends Run {
+	/**
+	 * Adds a racer with the param as the attribute
+	 * Cannot add a racer if a racer with the same bibNumber already Exists
+	 * @param bibNumber
+	 */
+	@Override
+	public void addRacer(int bibNumber){
+		Racer racer = new Racer(bibNumber);
+		if(waitQueue.contains(racer)||runningQueue.contains(racer)||endQueue.contains(racer)){
+			System.out.println("Cannot have more than one racer with the same bib number");
+			return;
+		}
+		waitQueue.add(racer);
+	}
+	/**
+	 * Gives start time to the first Racer in the queue
+	 * Adds them to the running queue indicating that the racer still needs an endtime
+	 */
+	@Override
+	public void setRacerStartTime(){
+		Racer headWait = waitQueue.poll();
+		headWait.setStartTime();
+		runningQueue.add(headWait);
+	}
+	/**
+	 * Gives end time  to the first Racer in the running queue
+	 * Adds the Racer to the endQueue indicating that the Racer is complete and finished
+	 */
+	@Override
+	public void setRacerEndTime() {
+		Racer headRunning = runningQueue.poll();
+		headRunning.setEndTime();
+		endQueue.add(headRunning);
+	}
+	/**
+	 * Sets the end time of a Racer that is running to DNF(-1)
+	 */
+	@Override
+	public void giveDnf() {
+		Racer headRunning = runningQueue.poll();
+		headRunning.setDnf();
+		endQueue.add(headRunning);
+	}
+	/**
+	 * Takes the first Racer out of the running queue and places them at the front of the waitQueue
+	 * Clears that Racers start time 
+	 */
+	@Override
+	public void cancel() {
+		Racer headRunning = runningQueue.poll();
+		headRunning.clearStartTime();
+		waitQueue.addFirst(headRunning);
+	}
+}
